refactor: introduce ConnectionManager interface to change connection easily
