+++ src/main/java/core/jdbc/ConnectionManager.java
@@ -1,31 +1,7 @@
 package core.jdbc;
 
-import org.apache.commons.dbcp2.BasicDataSource;
-
-import javax.sql.DataSource;
 import java.sql.Connection;
-import java.sql.SQLException;
 
-public class ConnectionManager {
-    private static final String DB_DRIVER = "org.h2.Driver";
-    private static final String DB_URL = "jdbc:h2:mem://localhost/~/jwp-jdbc;DB_CLOSE_DELAY=-1";
-    private static final String DB_USERNAME = "sa";
-    private static final String DB_PW = "";
-
-    public static DataSource getDataSource() {
-        BasicDataSource ds = new BasicDataSource();
-        ds.setDriverClassName(DB_DRIVER);
-        ds.setUrl(DB_URL);
-        ds.setUsername(DB_USERNAME);
-        ds.setPassword(DB_PW);
-        return ds;
-    }
-
-    public static Connection getConnection() {
-        try {
-            return getDataSource().getConnection();
-        } catch (SQLException e) {
-            throw new IllegalStateException(e);
-        }
-    }
+public interface ConnectionManager {
+    Connection getConnection();
 }
