+++ src/main/java/core/jdbc/DefaultConnectionManager.java
@@ -1,0 +1,36 @@
+package core.jdbc;
+
+import org.apache.commons.dbcp2.BasicDataSource;
+
+import javax.sql.DataSource;
+import java.sql.Connection;
+import java.sql.SQLException;
+
+public class DefaultConnectionManager implements ConnectionManager {
+    private static final String DB_DRIVER = "org.h2.Driver";
+    private static final String DB_URL = "jdbc:h2:mem://localhost/~/jwp-jdbc;DB_CLOSE_DELAY=-1";
+    private static final String DB_USERNAME = "sa";
+    private static final String DB_PW = "";
+
+    public static DataSource getDataSource() {
+        BasicDataSource ds = new BasicDataSource();
+        ds.setDriverClassName(DB_DRIVER);
+        ds.setUrl(DB_URL);
+        ds.setUsername(DB_USERNAME);
+        ds.setPassword(DB_PW);
+        return ds;
+    }
+
+    public static Connection establishConnection() {
+        try {
+            return getDataSource().getConnection();
+        } catch (SQLException e) {
+            throw new IllegalStateException(e);
+        }
+    }
+
+    @Override
+    public Connection getConnection() {
+        return establishConnection();
+    }
+}
