+++ src/main/java/next/dao/UserDao.java
@@ -1,5 +1,6 @@
 package next.dao;
 
+import core.jdbc.DefaultConnectionManager;
 import core.jdbc.JdbcTemplate;
 import core.jdbc.RowConverter;
 import next.model.User;
@@ -7,7 +8,7 @@ import next.model.User;
 import java.util.List;
 
 public class UserDao {
-    private static final JdbcTemplate jdbcTemplate = new JdbcTemplate();
+    private static final JdbcTemplate jdbcTemplate = new JdbcTemplate(new DefaultConnectionManager());
     private static final RowConverter<User> resultSetToUserConverter =
             resultSet -> new User(
                     resultSet.getString("userId"),
