+++ src/main/java/next/support/context/ContextLoaderListener.java
@@ -1,6 +1,6 @@
 package next.support.context;
 
-import core.jdbc.ConnectionManager;
+import core.jdbc.DefaultConnectionManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.core.io.ClassPathResource;
@@ -19,7 +19,7 @@ public class ContextLoaderListener implements ServletContextListener {
     public void contextInitialized(ServletContextEvent sce) {
         ResourceDatabasePopulator populator = new ResourceDatabasePopulator();
         populator.addScript(new ClassPathResource("jwp.sql"));
-        DatabasePopulatorUtils.execute(populator, ConnectionManager.getDataSource());
+        DatabasePopulatorUtils.execute(populator, DefaultConnectionManager.getDataSource());
 
         logger.info("Completed Load ServletContext!");
     }
